Export ContractAbi from types
